refactor(evaluators): shift evaluator code into the table package for future development
