+++ visitors.go
@@ -30,7 +30,7 @@ import (
 type BooleanExprVisitor[T any] interface {
 	VisitTrue() T
 	VisitFalse() T
-	VisitNot(childREsult T) T
+	VisitNot(childResult T) T
 	VisitAnd(left, right T) T
 	VisitOr(left, right T) T
 	VisitUnbound(UnboundPredicate) T
@@ -395,470 +395,3 @@ func (rewriteNotVisitor) VisitUnbound(pred UnboundPredicate) BooleanExpression {
 func (rewriteNotVisitor) VisitBound(pred BoundPredicate) BooleanExpression {
 	return pred
 }
-
-const (
-	rowsMightMatch, rowsMustMatch      = true, true
-	rowsCannotMatch, rowsMightNotMatch = false, false
-	inPredicateLimit                   = 200
-)
-
-// NewManifestEvaluator returns a function that can be used to evaluate whether a particular
-// manifest file has rows that might or might not match a given partition filter by using
-// the stats provided in the partitions (UpperBound/LowerBound/ContainsNull/ContainsNaN).
-func NewManifestEvaluator(spec PartitionSpec, schema *Schema, partitionFilter BooleanExpression, caseSensitive bool) (func(ManifestFile) (bool, error), error) {
-	partType := spec.PartitionType(schema)
-	partSchema := NewSchema(0, partType.FieldList...)
-	filter, err := RewriteNotExpr(partitionFilter)
-	if err != nil {
-		return nil, err
-	}
-
-	boundFilter, err := BindExpr(partSchema, filter, caseSensitive)
-	if err != nil {
-		return nil, err
-	}
-
-	return (&manifestEvalVisitor{partitionFilter: boundFilter}).Eval, nil
-}
-
-type manifestEvalVisitor struct {
-	partitionFields []FieldSummary
-	partitionFilter BooleanExpression
-}
-
-func (m *manifestEvalVisitor) Eval(manifest ManifestFile) (bool, error) {
-	if parts := manifest.Partitions(); len(parts) > 0 {
-		m.partitionFields = parts
-		return VisitExpr(m.partitionFilter, m)
-	}
-
-	return rowsMightMatch, nil
-}
-
-func allBoundCmp[T LiteralType](bound Literal, set Set[Literal], want int) bool {
-	val := bound.(TypedLiteral[T])
-	cmp := val.Comparator()
-
-	return set.All(func(e Literal) bool {
-		return cmp(val.Value(), e.(TypedLiteral[T]).Value()) == want
-	})
-}
-
-func allBoundCheck(bound Literal, set Set[Literal], want int) bool {
-	switch bound.Type().(type) {
-	case BooleanType:
-		return allBoundCmp[bool](bound, set, want)
-	case Int32Type:
-		return allBoundCmp[int32](bound, set, want)
-	case Int64Type:
-		return allBoundCmp[int64](bound, set, want)
-	case Float32Type:
-		return allBoundCmp[float32](bound, set, want)
-	case Float64Type:
-		return allBoundCmp[float64](bound, set, want)
-	case DateType:
-		return allBoundCmp[Date](bound, set, want)
-	case TimeType:
-		return allBoundCmp[Time](bound, set, want)
-	case TimestampType, TimestampTzType:
-		return allBoundCmp[Timestamp](bound, set, want)
-	case BinaryType, FixedType:
-		return allBoundCmp[[]byte](bound, set, want)
-	case StringType:
-		return allBoundCmp[string](bound, set, want)
-	case UUIDType:
-		return allBoundCmp[uuid.UUID](bound, set, want)
-	case DecimalType:
-		return allBoundCmp[Decimal](bound, set, want)
-	}
-	panic(ErrType)
-}
-
-func (m *manifestEvalVisitor) VisitIn(term BoundTerm, literals Set[Literal]) bool {
-	pos := term.Ref().Pos()
-	field := m.partitionFields[pos]
-
-	if field.LowerBound == nil {
-		return rowsCannotMatch
-	}
-
-	if literals.Len() > inPredicateLimit {
-		return rowsMightMatch
-	}
-
-	lower, err := LiteralFromBytes(term.Type(), *field.LowerBound)
-	if err != nil {
-		panic(err)
-	}
-
-	if allBoundCheck(lower, literals, 1) {
-		return rowsCannotMatch
-	}
-
-	if field.UpperBound != nil {
-		upper, err := LiteralFromBytes(term.Type(), *field.UpperBound)
-		if err != nil {
-			panic(err)
-		}
-
-		if allBoundCheck(upper, literals, -1) {
-			return rowsCannotMatch
-		}
-	}
-
-	return rowsMightMatch
-}
-
-func (m *manifestEvalVisitor) VisitNotIn(term BoundTerm, literals Set[Literal]) bool {
-	// because the bounds are not necessarily a min or max value, this cannot be answered using them
-	// notIn(col, {X, ...}) with (X, Y) doesn't guarantee that X is a value in col
-	return rowsMightMatch
-}
-
-func (m *manifestEvalVisitor) VisitIsNan(term BoundTerm) bool {
-	pos := term.Ref().Pos()
-	field := m.partitionFields[pos]
-
-	if field.ContainsNaN != nil && !*field.ContainsNaN {
-		return rowsCannotMatch
-	}
-
-	return rowsMightMatch
-}
-
-func (m *manifestEvalVisitor) VisitNotNan(term BoundTerm) bool {
-	pos := term.Ref().Pos()
-	field := m.partitionFields[pos]
-
-	if field.ContainsNaN != nil && *field.ContainsNaN && !field.ContainsNull && field.LowerBound == nil {
-		return rowsCannotMatch
-	}
-
-	return rowsMightMatch
-}
-
-func (m *manifestEvalVisitor) VisitIsNull(term BoundTerm) bool {
-	pos := term.Ref().Pos()
-	field := m.partitionFields[pos]
-
-	if !field.ContainsNull {
-		return rowsCannotMatch
-	}
-
-	return rowsMightMatch
-}
-
-func (m *manifestEvalVisitor) VisitNotNull(term BoundTerm) bool {
-	pos := term.Ref().Pos()
-	field := m.partitionFields[pos]
-
-	// ContainsNull encodes whether at least one partition value is null
-	// lowerBound is null if all partition values are null
-	allNull := field.ContainsNull && field.LowerBound == nil
-	if allNull && (term.Ref().Type().Equals(PrimitiveTypes.Float32) || term.Ref().Type().Equals(PrimitiveTypes.Float64)) {
-		// floating point types may include NaN values, which we check separately
-		// in case bounds don't include NaN values, ContainsNaN needsz to be checked
-		allNull = field.ContainsNaN != nil && !*field.ContainsNaN
-	}
-
-	if allNull {
-		return rowsCannotMatch
-	}
-
-	return rowsMightMatch
-}
-
-func getCmp[T LiteralType](b TypedLiteral[T]) func(Literal, Literal) int {
-	cmp := b.Comparator()
-	return func(l1, l2 Literal) int {
-		return cmp(l1.(TypedLiteral[T]).Value(), l2.(TypedLiteral[T]).Value())
-	}
-}
-
-func getCmpLiteral(boundary Literal) func(Literal, Literal) int {
-	switch l := boundary.(type) {
-	case TypedLiteral[bool]:
-		return getCmp(l)
-	case TypedLiteral[int32]:
-		return getCmp(l)
-	case TypedLiteral[int64]:
-		return getCmp(l)
-	case TypedLiteral[float32]:
-		return getCmp(l)
-	case TypedLiteral[float64]:
-		return getCmp(l)
-	case TypedLiteral[Date]:
-		return getCmp(l)
-	case TypedLiteral[Time]:
-		return getCmp(l)
-	case TypedLiteral[Timestamp]:
-		return getCmp(l)
-	case TypedLiteral[[]byte]:
-		return getCmp(l)
-	case TypedLiteral[string]:
-		return getCmp(l)
-	case TypedLiteral[uuid.UUID]:
-		return getCmp(l)
-	case TypedLiteral[Decimal]:
-		return getCmp(l)
-	}
-	panic(ErrType)
-}
-
-func (m *manifestEvalVisitor) VisitEqual(term BoundTerm, lit Literal) bool {
-	pos := term.Ref().Pos()
-	field := m.partitionFields[pos]
-
-	if field.LowerBound == nil || field.UpperBound == nil {
-		// values are all null and literal cannot contain null
-		return rowsCannotMatch
-	}
-
-	lower, err := LiteralFromBytes(term.Ref().Type(), *field.LowerBound)
-	if err != nil {
-		panic(err)
-	}
-
-	cmp := getCmpLiteral(lower)
-	if cmp(lower, lit) == 1 {
-		return rowsCannotMatch
-	}
-
-	upper, err := LiteralFromBytes(term.Ref().Type(), *field.UpperBound)
-	if err != nil {
-		panic(err)
-	}
-
-	if cmp(lit, upper) == 1 {
-		return rowsCannotMatch
-	}
-
-	return rowsMightMatch
-}
-
-func (m *manifestEvalVisitor) VisitNotEqual(term BoundTerm, lit Literal) bool {
-	// because bounds are not necessarily a min or max, this cannot be answered
-	// using them. notEq(col, X) with (X, Y) doesn't guarantee X is a value in col
-	return rowsMightMatch
-}
-
-func (m *manifestEvalVisitor) VisitGreaterEqual(term BoundTerm, lit Literal) bool {
-	pos := term.Ref().Pos()
-	field := m.partitionFields[pos]
-
-	if field.UpperBound == nil {
-		return rowsCannotMatch
-	}
-
-	upper, err := LiteralFromBytes(term.Ref().Type(), *field.UpperBound)
-	if err != nil {
-		panic(err)
-	}
-
-	if getCmpLiteral(upper)(lit, upper) == 1 {
-		return rowsCannotMatch
-	}
-
-	return rowsMightMatch
-}
-
-func (m *manifestEvalVisitor) VisitGreater(term BoundTerm, lit Literal) bool {
-	pos := term.Ref().Pos()
-	field := m.partitionFields[pos]
-
-	if field.UpperBound == nil {
-		return rowsCannotMatch
-	}
-
-	upper, err := LiteralFromBytes(term.Ref().Type(), *field.UpperBound)
-	if err != nil {
-		panic(err)
-	}
-
-	if getCmpLiteral(upper)(lit, upper) >= 0 {
-		return rowsCannotMatch
-	}
-
-	return rowsMightMatch
-}
-
-func (m *manifestEvalVisitor) VisitLessEqual(term BoundTerm, lit Literal) bool {
-	pos := term.Ref().Pos()
-	field := m.partitionFields[pos]
-
-	if field.LowerBound == nil {
-		return rowsCannotMatch
-	}
-
-	lower, err := LiteralFromBytes(term.Ref().Type(), *field.LowerBound)
-	if err != nil {
-		panic(err)
-	}
-
-	if getCmpLiteral(lower)(lit, lower) == -1 {
-		return rowsCannotMatch
-	}
-
-	return rowsMightMatch
-}
-
-func (m *manifestEvalVisitor) VisitLess(term BoundTerm, lit Literal) bool {
-	pos := term.Ref().Pos()
-	field := m.partitionFields[pos]
-
-	if field.LowerBound == nil {
-		return rowsCannotMatch
-	}
-
-	lower, err := LiteralFromBytes(term.Ref().Type(), *field.LowerBound)
-	if err != nil {
-		panic(err)
-	}
-
-	if getCmpLiteral(lower)(lit, lower) <= 0 {
-		return rowsCannotMatch
-	}
-
-	return rowsMightMatch
-}
-
-func (m *manifestEvalVisitor) VisitStartsWith(term BoundTerm, lit Literal) bool {
-	pos := term.Ref().Pos()
-	field := m.partitionFields[pos]
-
-	var prefix string
-	if val, ok := lit.(TypedLiteral[string]); ok {
-		prefix = val.Value()
-	} else {
-		prefix = string(lit.(TypedLiteral[[]byte]).Value())
-	}
-
-	lenPrefix := len(prefix)
-
-	if field.LowerBound == nil {
-		return rowsCannotMatch
-	}
-
-	lower, err := LiteralFromBytes(term.Ref().Type(), *field.LowerBound)
-	if err != nil {
-		panic(err)
-	}
-
-	// truncate lower bound so that it's length is not greater than the length of prefix
-	var v string
-	switch l := lower.(type) {
-	case TypedLiteral[string]:
-		v = l.Value()
-		if len(v) > lenPrefix {
-			v = v[:lenPrefix]
-		}
-	case TypedLiteral[[]byte]:
-		v = string(l.Value())
-		if len(v) > lenPrefix {
-			v = v[:lenPrefix]
-		}
-	}
-
-	if v > prefix {
-		return rowsCannotMatch
-	}
-
-	if field.UpperBound == nil {
-		return rowsCannotMatch
-	}
-
-	upper, err := LiteralFromBytes(term.Ref().Type(), *field.UpperBound)
-	if err != nil {
-		panic(err)
-	}
-
-	switch u := upper.(type) {
-	case TypedLiteral[string]:
-		v = u.Value()
-		if len(v) > lenPrefix {
-			v = v[:lenPrefix]
-		}
-	case TypedLiteral[[]byte]:
-		v = string(u.Value())
-		if len(v) > lenPrefix {
-			v = v[:lenPrefix]
-		}
-	}
-
-	if v < prefix {
-		return rowsCannotMatch
-	}
-
-	return rowsMightMatch
-}
-
-func (m *manifestEvalVisitor) VisitNotStartsWith(term BoundTerm, lit Literal) bool {
-	pos := term.Ref().Pos()
-	field := m.partitionFields[pos]
-
-	if field.ContainsNull || field.LowerBound == nil || field.UpperBound == nil {
-		return rowsMightMatch
-	}
-
-	// NotStartsWith will match unless ALL values must start with the prefix.
-	// this happens when the lower and upper bounds BOTH start with the prefix
-	lower, err := LiteralFromBytes(term.Ref().Type(), *field.LowerBound)
-	if err != nil {
-		panic(err)
-	}
-
-	upper, err := LiteralFromBytes(term.Ref().Type(), *field.UpperBound)
-	if err != nil {
-		panic(err)
-	}
-
-	var (
-		prefix, lowerBound, upperBound string
-	)
-	if val, ok := lit.(TypedLiteral[string]); ok {
-		prefix = val.Value()
-		lowerBound, upperBound = lower.(TypedLiteral[string]).Value(), upper.(TypedLiteral[string]).Value()
-	} else {
-		prefix = string(lit.(TypedLiteral[[]byte]).Value())
-		lowerBound = string(lower.(TypedLiteral[[]byte]).Value())
-		upperBound = string(upper.(TypedLiteral[[]byte]).Value())
-	}
-
-	lenPrefix := len(prefix)
-	if len(lowerBound) < lenPrefix {
-		return rowsMightMatch
-	}
-
-	if lowerBound[:lenPrefix] == prefix {
-		// if upper is shorter then upper can't start with the prefix
-		if len(upperBound) < lenPrefix {
-			return rowsMightMatch
-		}
-
-		if upperBound[:lenPrefix] == prefix {
-			return rowsCannotMatch
-		}
-	}
-
-	return rowsMightMatch
-}
-
-func (m *manifestEvalVisitor) VisitTrue() bool {
-	return rowsMightMatch
-}
-
-func (m *manifestEvalVisitor) VisitFalse() bool {
-	return rowsCannotMatch
-}
-
-func (m *manifestEvalVisitor) VisitUnbound(UnboundPredicate) bool {
-	panic("need bound predicate")
-}
-
-func (m *manifestEvalVisitor) VisitBound(pred BoundPredicate) bool {
-	return VisitBoundPredicate(pred, m)
-}
-
-func (m *manifestEvalVisitor) VisitNot(child bool) bool       { return !child }
-func (m *manifestEvalVisitor) VisitAnd(left, right bool) bool { return left && right }
-func (m *manifestEvalVisitor) VisitOr(left, right bool) bool  { return left || right }
